Using light/dark theme paparazzi tests.
