--- android/design-system/src/test/java/com/adammcneilly/pocketleague/android/designsystem/BasePaparazziTest.kt
@@ -1,49 +1,0 @@
-package com.adammcneilly.pocketleague.android.designsystem
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import app.cash.paparazzi.Paparazzi
-import com.adammcneilly.pocketleague.android.designsystem.theme.PocketLeagueTheme
-import com.google.testing.junit.testparameterinjector.TestParameter
-import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import org.junit.Rule
-import org.junit.runner.RunWith
-
-@RunWith(TestParameterInjector::class)
-open class BasePaparazziTest {
-
-    @get:Rule
-    val paparazzi = Paparazzi()
-
-    @TestParameter
-    val useDarkTheme: Boolean = false
-
-    open fun snapshotScreen(
-        useDarkTheme: Boolean = this.useDarkTheme,
-        content: @Composable () -> Unit,
-    ) {
-        paparazzi.snapshot {
-            PocketLeagueTheme(
-                useDarkTheme = useDarkTheme,
-            ) {
-                Surface(
-                    modifier = Modifier
-                        .fillMaxSize(),
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .padding(16.dp),
-                    ) {
-                        content()
-                    }
-                }
-            }
-        }
-    }
-}
