+++ android/design-system/src/test/java/com/adammcneilly/pocketleague/android/designsystem/components/EmptyStateCardPaparazziTest.kt
@@ -3,14 +3,28 @@ package com.adammcneilly.pocketleague.android.designsystem.components
 import androidx.compose.foundation.layout.padding
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.adammcneilly.pocketleague.android.designsystem.BasePaparazziTest
+import app.cash.paparazzi.Paparazzi
+import com.adammcneilly.pocketleague.android.designsystem.snapshotScreen
+import com.google.testing.junit.testparameterinjector.TestParameter
+import com.google.testing.junit.testparameterinjector.TestParameterInjector
+import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
 
-class EmptyStateCardPaparazziTest : BasePaparazziTest() {
+@RunWith(TestParameterInjector::class)
+class EmptyStateCardPaparazziTest {
+
+    @get:Rule
+    val paparazzi = Paparazzi()
+
+    @TestParameter
+    val useDarkTheme: Boolean = false
 
     @Test
     fun renderCard() {
-        snapshotScreen {
+        paparazzi.snapshotScreen(
+            useDarkTheme = useDarkTheme,
+        ) {
             EmptyStateCard(
                 text = "Demo Empty State Card",
                 textModifier = Modifier.padding(32.dp),
