+++ android/design-system/src/test/java/com/adammcneilly/pocketleague/android/designsystem/components/InlineIconTextPaparazziTest.kt
@@ -4,16 +4,26 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.EmojiEvents
 import app.cash.paparazzi.Paparazzi
 import com.adammcneilly.pocketleague.android.designsystem.snapshotScreen
+import com.google.testing.junit.testparameterinjector.TestParameter
+import com.google.testing.junit.testparameterinjector.TestParameterInjector
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
 
+@RunWith(TestParameterInjector::class)
 class InlineIconTextPaparazziTest {
+
     @get:Rule
     val paparazzi = Paparazzi()
 
+    @TestParameter
+    val useDarkTheme: Boolean = false
+
     @Test
     fun renderWithoutIcon() {
-        paparazzi.snapshotScreen {
+        paparazzi.snapshotScreen(
+            useDarkTheme = useDarkTheme,
+        ) {
             InlineIconText(
                 text = "Inline Icon Text",
                 icon = Icons.Default.EmojiEvents,
@@ -24,7 +34,9 @@ class InlineIconTextPaparazziTest {
 
     @Test
     fun renderLeadingIcon() {
-        paparazzi.snapshotScreen {
+        paparazzi.snapshotScreen(
+            useDarkTheme = useDarkTheme,
+        ) {
             InlineIconText(
                 text = "Inline Icon Text",
                 icon = Icons.Default.EmojiEvents,
@@ -35,7 +47,9 @@ class InlineIconTextPaparazziTest {
 
     @Test
     fun renderTrailingIcon() {
-        paparazzi.snapshotScreen {
+        paparazzi.snapshotScreen(
+            useDarkTheme = useDarkTheme,
+        ) {
             InlineIconText(
                 text = "Inline Icon Text",
                 icon = Icons.Default.EmojiEvents,
