+++ android/design-system/src/test/snapshots/images/com.adammcneilly.pocketleague.android.designsystem.components_InlineIconTextPaparazziTest_renderWithoutIcon[useDarkTheme=false].png
