+++ android/design-system/src/test/snapshots/images/com.adammcneilly.pocketleague.android.designsystem.components_MatchCardPaparazziTest_renderBlueTeamWinner[useDarkTheme=false].png
