+++ android/design-system/src/test/snapshots/images/com.adammcneilly.pocketleague.android.designsystem.components_MatchCardPaparazziTest_renderPlaceholder[useDarkTheme=false].png
